+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.0.2
 
 - Add `Rdmm::Responses::BaseResponse#total_count`
+- Add `Rdmm::Client#list_series`
+- Rename `#series_list` to `#series`
 
 ## 0.0.1